// lib/types.ts
export type ActionId =
  | 'brush' | 'smash' | 'cut' | 'separate' | 'dig'
  | 'carve' | 'scrape' | 'grind' | 'pull' | 'twist' | 'tie' | 'stretch' | 'mix' | 'shape'
  | 'pour' | 'heat' | 'cool' | 'dry' | 'burn' | 'hammer' | 'split' | 'press' | 'saw'
  | 'chisel' | 'polish';